Added
This script is used to add distance columns to a given pandas.DataFrame.
In my previous computation of v_min, I have forgotten to include the distance columns for some of the catalogues, so
I wrote this script to redo the MCMC sampling for distances without the need to run the whole computation script again.