Minor updates (temporary)
Since in the new combined catalogues, RA and Dec for Gaia sources became ra_gaia and dec_gaia, I added these lines that check for "ra" and "dec" in the input DataFrame, and rename them if there are not. This change is made to make the code compatible with old and new combined catalogues, but may just be a temporary change.